last

+++ ft_putstr_fd.c
@@ -2,7 +2,7 @@
 
 void	ft_putstr_fd(char *s, int fd)
 {
-	unsigned long	i;
+	size_t	i;
 
 	i = ft_strlen(s);
 	write(fd, s, i);
